+++ services/collaboration-service/package.json
@@ -7,6 +7,7 @@
     "swagger-autogen": "node swagger.js"
   },
   "dependencies": {
+    "@types/express": "^4.17.17",
     "body-parser": "^1.20.2",
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
@@ -16,6 +17,7 @@
     "socket.io": "^4.7.2",
     "swagger-autogen": "^2.23.5",
     "swagger-ui-express": "^5.0.0",
+    "typescript": "^5.2.2",
     "uuid": "^9.0.1"
   },
   "devDependencies": {
