+++ services/collaboration-service/swagger.ts
@@ -1,3 +1,4 @@
+// @ts-expect-error TS(2580): Cannot find name 'require'. Do you need to install... Remove this comment to see the full error message
 const swaggerAutogen = require("swagger-autogen")();
 
 const doc = {
